+++ src/commands/openAIChatCommand.ts
@@ -19,7 +19,8 @@ import {
 import { IgnoreFileManager } from "../realtimeScanners/common/ignoreFileManager";
 import { OssScannerService } from "../realtimeScanners/scanners/oss/ossScannerService";
 import { SecretsScannerService } from "../realtimeScanners/scanners/secrets/secretsScannerService";
-import path from "path";
+
+import { cx } from "../cx";
 
 
 export class CopilotChatCommand {
@@ -127,14 +128,30 @@ export class CopilotChatCommand {
         await vscode.commands.executeCommand(constants.copilotChatOpenWithQueryCommand, { query: `${question}` });
     }
 
+    private logUserEvent(EventType: string, subType: string, item: HoverData | SecretsHoverData | AscaHoverData | ContainersHoverData): void {
+        const isSecrets = isSecretsHoverData(item);
+        const engine = isSecrets ? constants.secretsScannerEngineName : constants.ossRealtimeScannerEngineName;
+        let problemSeverity: string | undefined;
+        if (isSecrets) {
+            problemSeverity = item.severity;
+        } else if (isAscaHoverData(item)) {
+            problemSeverity = item.severity;
+        } else {
+            problemSeverity = (item as HoverData | ContainersHoverData).status;
+        }
+        cx.setUserEventDataForLogs(EventType, subType, engine, problemSeverity);
+    }
 
 
 
     public registerCopilotChatCommand() {
         this.context.subscriptions.push(
             vscode.commands.registerCommand(commands.openAIChat, async (item: HoverData | SecretsHoverData | AscaHoverData | ContainersHoverData) => {
+                this.logUserEvent("click", constants.openAIChat, item);
+
+                const isSecrets = isSecretsHoverData(item);
                 let question = '';
-                if (isSecretsHoverData(item)) {
+                if (isSecrets) {
                     question = SECRET_REMEDIATION_PROMPT(item.title, item.description, item.severity);
                 } else if (isAscaHoverData(item)) {
                     question = ASCA_REMEDIATION_PROMPT(item.ruleName, item.description, item.severity, item.remediationAdvise);
@@ -153,9 +170,12 @@ export class CopilotChatCommand {
         );
         this.context.subscriptions.push(
             vscode.commands.registerCommand(commands.viewDetails, async (item: HoverData | SecretsHoverData | AscaHoverData | ContainersHoverData) => {
+                this.logUserEvent("click", constants.viewDetails, item);
+
+                const isSecrets = isSecretsHoverData(item);
                 let question = '';
 
-                if (isSecretsHoverData(item)) {
+                if (isSecrets) {
                     question = SECRETS_EXPLANATION_PROMPT(item.title, item.description, item.severity);
                 } else if (isAscaHoverData(item)) {
                     question = ASCA_EXPLANATION_PROMPT(item.ruleName, item.description, item.severity);