+++ src/cx/cx.ts
@@ -6,7 +6,7 @@ import CxProject from "@checkmarxdev/ast-cli-javascript-wrapper/dist/main/projec
 import CxCodeBashing from "@checkmarxdev/ast-cli-javascript-wrapper/dist/main/codebashing/CxCodeBashing";
 import { CxConfig } from "@checkmarxdev/ast-cli-javascript-wrapper/dist/main/wrapper/CxConfig";
 import { constants } from "../utils/common/constants";
-import { getFilePath, getResultsFilePath } from "../utils/utils";
+import { getFilePath, getResultsFilePath, isCursorIDE } from "../utils/utils";
 import { SastNode } from "../models/sastNode";
 import AstError from "../exceptions/AstError";
 import { CxParamType } from "@checkmarxdev/ast-cli-javascript-wrapper/dist/main/wrapper/CxParamType";
@@ -738,5 +738,14 @@ export class Cx implements CxPlatform {
     }
 
     return r;
+  }
+
+  async setUserEventDataForLogs(eventType: string, subType:string, engine: string, problemSeverity: string) {
+    const config = await this.getAstConfiguration();
+    const cx = new CxWrapper(config);
+    const aiProvider = isCursorIDE() ? "Cursor" : "Copilot";
+    const agent = isCursorIDE() ? "Cursor" : constants.vsCodeAgent;
+
+    cx.telemetryAIEvent(aiProvider, agent, eventType, subType, engine, problemSeverity);
   }
 }