+++ src/cx/cxMock.ts
@@ -1650,4 +1650,6 @@ export class CxMock implements CxPlatform {
       ],
     });
   }
+
+  setUserEventDataForLogs(): void {}
 }