+++ src/utils/common/commands.ts
@@ -119,9 +119,10 @@ export const commands = {
 
   openSettings: "workbench.action.openSettings",
   openSettingsArgs: `@ext:checkmarx.${constants.extensionName}`,
-  openAIChat: `${constants.extensionName}.fixWithAIChat`,
-  viewDetails: `${constants.extensionName}.viewPackageDetails`,
+  openAIChat: `${constants.extensionName}.${constants.openAIChat}`,
+  viewDetails: `${constants.extensionName}.${constants.viewDetails}`,
   ignorePackage: `${constants.extensionName}.ignorePackage`,
   IgnoreAll: `${constants.extensionName}.ignoreAll`,
   openIgnoredView: `${constants.extensionName}.openIgnoredView`,
+
 };